# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "192.168.1.41:1323",
    "basePath": "/",
    "paths": {
...
        "/onboarding/doctor": {
            "post": {
                "description": "Create a new Doctor user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Doctor.Create",
                "parameters": [
                    {
                        "description": "Doctor",
                        "name": "Doctor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.formDoctor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.doctOut"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/onboarding/patient": {
            "post": {
                "description": "Create a new Patient user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Patient.Create",
                "parameters": [
                    {
                        "description": "Doctor",
                        "name": "Doctor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.formPatient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.formResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/patients/{pati_id}": {
            "get": {
                "description": "Get patient data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Patient.Get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Context to return",
                        "name": "context",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Patient ID",
                        "name": "pati_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.getResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Patient update with patient data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Patient.Update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Context to return",
                        "name": "context",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Patient ID",
                        "name": "pati_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update patient data",
                        "name": "patient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/user.Patient"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.getResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/signup/{id}": {
            "get": {
                "description": "get string by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Show a account",
                "operationId": "get-string-by-int",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/handler.User"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "account name"
                },
                "photo_urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "http://test/image/1.jpg",
                        "http://test/image/2.jpg"
                    ]
                }
            }
        },
...
        "handler.detailError": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string"
                },
                "extendedHelp": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "locationType": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "sendReport": {
                    "type": "string"
                }
            }
        },
        "handler.doctOut": {
            "type": "object",
            "properties": {
                "item": {
                    "type": "object",
                    "$ref": "#/definitions/user.Doctor"
                },
                "kind": {
                    "type": "string",
                    "example": "Doctor"
                },
                "lang": {
                    "description": "Indicates the language of the rest of the properties in this object (BCP 47)",
                    "type": "string",
                    "example": "pt-br"
                }
            }
        },
        "handler.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "$ref": "#/definitions/handler.generalError"
                }
            }
        },
...
        "handler.formDoctor": {
            "type": "object",
            "properties": {
                "crm": {
                    "type": "string"
                },
                "documents": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "info": {
                    "type": "object",
                    "$ref": "#/definitions/user.Info"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "handler.formPatient": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "cpf": {
                    "type": "string"
                },
                "documents": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "info": {
                    "type": "object",
                    "$ref": "#/definitions/user.Info"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "rg": {
                    "type": "string"
                }
            }
        },
        "handler.formResponse": {
            "type": "object",
            "properties": {
                "item": {
                    "type": "object",
                    "$ref": "#/definitions/handler.response"
                },
                "kind": {
                    "type": "string"
                },
                "lang": {
                    "description": "Indicates the language of the rest of the properties in this object (BCP 47)",
                    "type": "string",
                    "example": "pt-br"
                }
            }
        },
...
        "handler.generalError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.detailError"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.getResponse": {
            "type": "object",
            "properties": {
                "item": {
                    "type": "object",
                    "$ref": "#/definitions/handler.responseGet"
                },
                "kind": {
                    "type": "string",
                    "example": "Patient"
                },
                "lang": {
                    "description": "Indicates the language of the rest of the properties in this object (BCP 47)",
                    "type": "string",
                    "example": "pt-br"
                }
            }
        },
...
        "handler.response": {
            "type": "object",
            "properties": {
                "jwt": {
                    "type": "string",
                    "example": "wqeoifjweoifjwef.afoj3204jfdkjf0wjf0wefj0w9fjf..."
                },
                "user": {
                    "type": "object",
                    "$ref": "#/definitions/user.Patient"
                }
            }
        },
        "handler.responseGet": {
            "type": "object",
            "properties": {
                "user": {
                    "type": "object",
                    "$ref": "#/definitions/user.Patient"
                }
            }
        },
...
        "user.Doctor": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "crm": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "doctID": {
                    "type": "string"
                },
                "documents": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "info": {
                    "type": "object",
                    "$ref": "#/definitions/user.Info"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "object",
                    "$ref": "#/definitions/user.Role"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "user.Info": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
...
        "user.Patient": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "cpf": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "documents": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "info": {
                    "type": "object",
                    "$ref": "#/definitions/user.Info"
                },
                "name": {
                    "type": "string"
                },
                "patiID": {
                    "type": "string"
                },
                "rg": {
                    "type": "string"
                },
                "role": {
                    "type": "object",
                    "$ref": "#/definitions/user.Role"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
...
        "user.Role": {
            "type": "object"
        },
...
    }
}